all done

+++ src/controllers/notesController.js
@@ -3,10 +3,11 @@ import createHttpError from 'http-errors';
 
 export const getAllNotes = async (req, res) => {
   const { page = 1, perPage = 10, tag, search } = req.query;
+  const userId = req.user._id;
 
   const skip = (page - 1) * perPage;
 
-  const noteQuery = Note.find();
+  const noteQuery = Note.find({ userId });
 
   if (tag) {
     noteQuery.where('tag').equals(tag);
@@ -28,7 +29,9 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const userId = req.user._id;
+
+  const note = await Note.findById({ _id: noteId, userId });
 
   if (!note) {
     next(createHttpError(404, 'Note not found'));
@@ -38,13 +41,16 @@ export const getNoteById = async (req, res, next) => {
 };
 
 export const createNote = async (req, res) => {
-  const note = await Note.create(req.body);
+  const note = await Note.create({ ...req.body, userId: req.user._id });
+
   res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const userId = req.user._id;
+
+  const note = await Note.findOneAndDelete({ _id: noteId, userId });
 
   if (!note) {
     next(createHttpError(404, 'Note not found'));
@@ -56,7 +62,9 @@ export const deleteNote = async (req, res, next) => {
 
 export const updateNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndUpdate({ _id: noteId }, req.body, {
+  const userId = req.user._id;
+
+  const note = await Note.findOneAndUpdate({ _id: noteId, userId }, req.body, {
     new: true,
   });
